♻️  refactor: Get data of festival from API in festival signup + refacto format date in utils + Date in french
