+++ Projet_IOS_Ig4/Festival/Views/SignupFestivalView.swift
@@ -15,6 +15,7 @@ struct Poste {
 }
 
 struct SignupFestivalView: View {
+    @StateObject var festivalViewModel = FestivalViewModel()
     @State private var teeShirtSize: String = "XS"
     @State private var isVegetarian: Bool = false
     let festivalName: String
@@ -38,85 +39,93 @@ struct SignupFestivalView: View {
     let columnWidths: [CGFloat] = [150, 100]
     
     var body: some View {
-        VStack(alignment: .leading, spacing: 20) {
-            // Festival information
-            Text(festivalName)
-                .font(.title)
-                .fontWeight(.bold)
-            Text("Du \(startDate) au \(endDate)")
-            
-            // Tee shirt size selection
-            HStack() {
-                Text("Taille de t-shirt")
-                Spacer()
-                Picker("Tee Shirt Size", selection: $teeShirtSize) {
-                    Text("XS").tag("XS")
-                    Text("S").tag("S")
-                    Text("M").tag("M")
-                    Text("L").tag("L")
-                    Text("XL").tag("XL")
-                }
-                .pickerStyle(MenuPickerStyle())
-            }
-                        
-            // Vegetarian selection
-            Toggle("Vegetarien", isOn: $isVegetarian)
-            
-            // Postes selection
-            Text("Choix des postes")
-                .font(.title2)
-                .fontWeight(.bold)
-            Text("01/01/2024")
-                .font(.title3)
-                .fontWeight(.bold)
-            ScrollView(.horizontal) {
-                VStack {
-                    Text("") // Empty column
-                        .frame(width: columnWidths[0])
-                    
-                    // Header row with creneaux
-                    HStack {
-                        Spacer()
-                        ForEach(creneaux, id: \.id) { creneau in
-                            Text("\(creneau.timeStart)-\(creneau.timeEnd)")
-                                .font(.headline)
-                                .frame(width: columnWidths[1], alignment: .center)
-                        }
+        if let festival = festivalViewModel.festival {
+            VStack(alignment: .leading, spacing: 20) {
+                // Festival information
+                Text(festival.name)
+                    .font(.title)
+                    .fontWeight(.bold)
+                Text("Du \(DateUtils.formatDate(festival.dateDebut)!) au \(DateUtils.formatDate(festival.dateFin)!)")
+                
+                // Tee shirt size selection
+                HStack() {
+                    Text("Taille de t-shirt")
+                    Spacer()
+                    Picker("Tee Shirt Size", selection: $teeShirtSize) {
+                        Text("XS").tag("XS")
+                        Text("S").tag("S")
+                        Text("M").tag("M")
+                        Text("L").tag("L")
+                        Text("XL").tag("XL")
                     }
-                    
-                    // Data rows
-                    ForEach(postes, id: \.id) { poste in
+                    .pickerStyle(MenuPickerStyle())
+                }
+                
+                // Vegetarian selection
+                Toggle("Vegetarien", isOn: $isVegetarian)
+                
+                // Postes selection
+                Text("Choix des postes")
+                    .font(.title2)
+                    .fontWeight(.bold)
+                Text("01/01/2024")
+                    .font(.title3)
+                    .fontWeight(.bold)
+                ScrollView(.horizontal) {
+                    VStack {
+                        Text("") // Empty column
+                            .frame(width: columnWidths[0])
+                        
+                        // Header row with creneaux
                         HStack {
-                            Text(poste.name)
-                                .padding()
-                                .frame(width: columnWidths[0]) // Width for poste name
-                            
                             Spacer()
-                            
                             ForEach(creneaux, id: \.id) { creneau in
-                                CircularProgressView(current: 7, max: poste.capacity)
+                                Text("\(creneau.timeStart)-\(creneau.timeEnd)")
+                                    .font(.headline)
+                                    .frame(width: columnWidths[1], alignment: .center)
+                            }
+                        }
+                        
+                        // Data rows
+                        ForEach(postes, id: \.id) { poste in
+                            HStack {
+                                Text(poste.name)
                                     .padding()
-                                    .frame(width: columnWidths[1]) // Width for circular progress bar cell
+                                    .frame(width: columnWidths[0]) // Width for poste name
                                 
                                 Spacer()
+                                
+                                ForEach(creneaux, id: \.id) { creneau in
+                                    CircularProgressView(current: 7, max: poste.capacity)
+                                        .padding()
+                                        .frame(width: columnWidths[1]) // Width for circular progress bar cell
+                                    
+                                    Spacer()
+                                }
                             }
                         }
                     }
                 }
+                // Signup button
+                Button(action: {
+                    print("Signup button tapped")
+                }) {
+                    Text("Je m'inscris au festival")
+                        .padding()
+                        .frame(maxWidth: .infinity)
+                        .background(Color.blue)
+                        .foregroundColor(.white)
+                        .cornerRadius(10)
+                }
             }
-            // Signup button
-            Button(action: {
-                print("Signup button tapped")
-            }) {
-                Text("Je m'inscris au festival")
-                    .padding()
-                    .frame(maxWidth: .infinity)
-                    .background(Color.blue)
-                    .foregroundColor(.white)
-                    .cornerRadius(10)
-            }
+            .padding()
+        } else {
+            ProgressView("Chargement du festival...")
+                .progressViewStyle(CircularProgressViewStyle())
+                .onAppear {
+                    festivalViewModel.fetchFestival()
+                }
         }
-        .padding()
     }
 }
